healthd: fix offline charging.
-I implemented an lpm.rc so I could isolate the charging
mode initialization and simplify diagnostics. So, instead of
parsing init and all it's imports, during charging mode, we
are just dealing with lpm.rc. This may not be needed now
that I have it working but I was getting "reboots to recovery"
when using "init.rc" for charging mode.

-Got adb working in charging mode for diagnostics.

-Wrote to /sys/devices/virtual/graphics/fb0/mode fix the
blank screen issue.

-Set pixel format to BGRA_8888. Not sure if this is necessary
but it works. The TWRP we are currently using, uses this format.

-Need to disable adb and get all the diag stuff out of here..