patches: add fix for hw keystore (also fixes fingerprint)
-It seems that there are a couple "purposes" for
algorithm 0 that are not supported in Hisi's keymaster. So
re are just returning OK as if does. This could cause problems
if these are actually used somewhere but we will deal
with that when and if it comes up...

-Diagnostics patch is probably not needed but I wanted to
save it along with the keystore patch. Will be removed later.