Screencast:
-Screencast does not work with out hw avc encoder but it works
find on google's avc encoder but because our HW codecs are first
in the Media codec list. Under most circumstances this would be
fine. After all most of the time you want to use the HW codecs
but for this case we need to able to specify which codec to use.

-Apparently, the screenrecord command and the Screencast app
act independently of one another so there is a fix for both here.