recovery: installation checks
-Up till now installation has been a little complicated and
many users do not bother to follow the instructions. With this
we will make it easier. So here are the checks in order:

1. If they are not using Meticulus Development's TWRP - abort install
2. If they are trying to install over EMUI 4.1 - ask if it's
OK to format /data. If yes then we format it for them. If not - abort
install.
3. After a successful installation and if installing over EMUI 4.1,
TWRP is in a defunct state. Ask to reboot to recovery to switch
TWRP to CM (blue) mode.