libcopybit_wrapper: don't do debug dumps.
-fill_color is my crazy little function. Because of the way
"hw" blobs work, function names are less important that a
functions position is a stack. This functions name would be
more acurate if it was "create_hwc_channel" and I suspect
that this function is supposed to work with the hwcomposer.
However, we are not using the stock HWC and so this function fails
and to avoid a crash we just nullify it. I must have forgotten
all that when I added debug code. The function prototype is
most likely incorrect.