vendor_symlink: use existing vendor
-The idea here is to symlink in what works from vendor
to /system/vendor. We mount the vendor partition @ /hwvendor.
This way we can, for instance, simply not link in the stock
hwcomposwer which does not work for us and we don't have to
modify the vendor partition at all.
-A further benefit might be that each variant can keep
the ril blobs that was installed with their stock. Hopefully
this will eliminate some ril issues.