libhuawei_init: Fix up modem prop code.
-This gave me alot of trouble because the modem prop code had
a rare bug. When using 'getline' if you set the 'char *' to NULL
and 'size_t' to 0. 'getline' is will create a 'char *' the exact
'size_t' of the readline. However in the first condition of
the loop, in which 'continue' is used, linebuf is not
reset to NULL. So, in the next iter of the loop, crash.

-So, now we just allocate a buffer the max size of property.

-Simplify property spliting using 'strtok'

-NOTE: My modemid is not in my phone.prop?