TWRP: Fix stock decryption + Magisk
-In stock recovery, the file /res/native_packages.xml
is absolutely neccessary for teecd, volisnotd and stock
/data decrption. However, "PRODUCT_COPY_FILES" will not
copy a file with an xml extenstion unless it is "well-formed"
xml, which this file is not. So we rename the file
'native_packages.bin' but also, when trying to copy this
file to /recovery/root/res. It gets deleted by something
in build. So we copy it to /recovery/root/system/etc
and I use my fstabSwitcher to symlink to /res/native_packages.xml
at runtime.

-Swap out SuperSU for Magisk.